extracted db connection class
